requirements.txt
streamlit==1.35.0
sentence-transformers==2.7.0
torch==2.3.0
qdrant-client==1.9.1
neo4j==5.20.0
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
datasets==2.19.0
huggingface-hub==0.23.0
more-itertools==10.3.0
requests==2.32.3
tqdm==4.66.4
python-dotenv==1.0.1
